chore: mos-gateway and playout-gateway connection refactoring
