+++ packages/server-core-integration/examples/client.ts
@@ -1,6 +1,7 @@
 import { protectString } from '@sofie-automation/shared-lib/dist/lib/protectedString'
 import { StatusCode } from '@sofie-automation/shared-lib/dist/lib/status'
 import {
+	PERIPHERAL_SUBTYPE_PROCESS,
 	PeripheralDeviceCategory,
 	PeripheralDeviceType,
 } from '@sofie-automation/shared-lib/dist/peripheralDevice/peripheralDeviceAPI'
@@ -11,7 +12,11 @@ const core = new CoreConnection({
 	deviceToken: 'abcd',
 	deviceCategory: PeripheralDeviceCategory.PLAYOUT,
 	deviceType: PeripheralDeviceType.PLAYOUT,
+	deviceSubType: PERIPHERAL_SUBTYPE_PROCESS,
 	deviceName: 'Jest test framework',
+	documentationUrl: 'http://example.com',
+	versions: {},
+	configManifest: {} as any,
 })
 
 core.onConnectionChanged((connected) => {
